Module psl_core::datamodel_connector

The interface implemented by connectors for Prisma schema validation and interpretation.

Re-exports

pub use self::capabilities::ConnectorCapabilities;

pub use self::capabilities::ConnectorCapability;

pub use self::completions::format_completion_docs;

Modules

capabilities

Connector capabilities

completions

Connector completions

constraint_names

Constraint name defaults.

walker_ext_traits

Extensions for parser database walkers with context from the connector.

Structs

EmptyDatamodelConnector

A Connector implementor meant to be used as a default when no datasource is defined.

NativeTypeConstructor

Represents an available native type.

NativeTypeInstance

Enums

ConstraintScope

A scope where a constraint name must be unique.

ConstraintType

Flavour

RelationMode

Defines the part of the stack where referential actions are handled.

StringFilter

Available filters for a given String scalar field.

Constants

EXTENSIONS_KEY

Traits

Connector

The datamodel connector API.

NativeTypeArguments