Struct psl_core::js_connector::JsConnector

pub struct JsConnector {
    pub flavour: Flavour,
    pub canonical_connector: &'static dyn Connector,
    pub provider_name: &'static str,
    pub name: &'static str,
    pub allowed_protocols: Option<&'static [&'static str]>,
}

JsConnector represents a type of connector that is implemented partially in javascript and used from rust through the js-connectors crate

Rather than a unit struct per individual connector, like we have for the rest of the builtin connectors, we have a single struct which state represents the features that vary in this connector with respect to a cannonical connector for the flavour of SQL the particular JsConnector speaks.

For example, the planetscale serverless connector is compatible with MySQL, so it reuses the builtin MySQL connector (the cannonical for the MySQL flavour) for most of its features.

Fields
flavour: Flavourcanonical_connector: &'static dyn Connectorprovider_name: &'static strname: &'static strallowed_protocols: Option<&'static [&'static str]>
Implementations

impl JsConnector

pub fn is_provider(name: &str) -> bool

Returns true if the given name is a valid provider name for a JsConnector. We use the convention that if a provider starts with ´@prisma/´ (ex. ´@prisma/planetscale´) then its a provider for a JS connector.