Crate psl_core

PSL core

This crate is the entry point for the core of the PSL implementation. It exposes the Connector trait and relies on its implementors, but it is itself decoupled from any specific connector.

Re-exports

pub use diagnostics;

pub use parser_database;

pub use schema_ast;

Modules

builtin_connectors

datamodel_connector

The interface implemented by connectors for Prisma schema validation and interpretation.

mcf

mcf: Turns a collection of configuration::Datasource and configuration::Generator into a JSON representation. This is the get_config() representation.

Macros

native_type_definition

Structs

Configuration

Datasource

a datasource from the prisma schema.

DatasourceConnectorData

Generator

StringFromEnvVar

Either an env var or a string literal.

ValidatedSchema

Enums

GeneratorConfigValue

PreviewFeature

Constants

ALL_PREVIEW_FEATURES

Generator preview features (alphabetically sorted)

Functions

is_reserved_type_name

Is this a valid type name for the Prisma Client API?

parse_configuration

Loads all configuration blocks from a datamodel using the built-in source definitions.

parse_without_validation

Retrieves a Prisma schema without validating it. You should only use this method when actually validating the schema is too expensive computationally or in terms of bundle size (e.g., for query-engine-wasm).

reformat

Returns either the reformatted schema, or the original input if we can’t reformat. This happens if and only if the source does not parse to a well formed AST.

set_config_dir

validate

The most general API for dealing with Prisma schemas. It accumulates what analysis and validation information it can, and returns it along with any error and warning diagnostics.

Type Aliases

ConnectorRegistry

The collection of all available connectors.

PreviewFeatures

A set of preview features.