Crate psl_core

PSL core

This crate is the entry point for the core of the PSL implementation. It exposes the Connector trait and relies on its implementors, but it is itself decoupled from any specific connector.

Re-exports

pub use diagnostics;

pub use parser_database;

pub use schema_ast;

Modules

datamodel_connector

The interface implemented by connectors for Prisma schema validation and interpretation.

mcf

mcf: Turns a collection of configuration::Datasource and configuration::Generator into a JSON representation. This is the get_config() representation.

Structs

Configuration

Datasource

a datasource from the prisma schema.

DatasourceConnectorData

Generator

StringFromEnvVar

Either an env var or a string literal.

ValidatedSchema

Enums

PreviewFeature

Constants

ALL_PREVIEW_FEATURES

Generator preview features

Functions

is_reserved_type_name

Is this a valid type name for the Prisma client API?

parse_configuration

Loads all configuration blocks from a datamodel using the built-in source definitions.

reformat

Returns either the reformatted schema, or the original input if we can’t reformat. This happens if and only if the source does not parse to a well formed AST.

validate

The most general API for dealing with Prisma schemas. It accumulates what analysis and validation information it can, and returns it along with any error and warning diagnostics.

Type Definitions

ConnectorRegistry

The collection of all available connectors.

PreviewFeatures

A set of preview features.